+++ spec/requests/articles_spec.rb
@@ -8,7 +8,7 @@ describe "Articles management", :type => :request, :order => :defined do
 
     it "should not create a version" do
       expect(PaperTrail).to be_enabled_for_controller
-      expect { post(articles_path, valid_params) }.to_not change(PaperTrail::Version, :count)
+      expect { post articles_path, params_wrapper(valid_params) }.to_not change(PaperTrail::Version, :count)
     end
 
     it "should not leak the state of the `PaperTrail.enabled_for_controller?` into the next test" do
@@ -21,7 +21,7 @@ describe "Articles management", :type => :request, :order => :defined do
 
     context "`current_user` method returns a `String`" do
       it "should set that value as the `whodunnit`" do
-        expect { post articles_path, valid_params }.to change(PaperTrail::Version, :count).by(1)
+        expect { post articles_path, params_wrapper(valid_params) }.to change(PaperTrail::Version, :count).by(1)
         expect(article.title).to eq('Doh')
         expect(article.versions.last.whodunnit).to eq('foobar')
       end